pyproject.toml
[tool.poetry]
name = "referral"
version = "0.1.0"
description = ""
authors = ["Dmitrii Ivanov <dmitriy.site1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
